CHANGE: passenger_max_wait is now a uint32 value, and the default is increased to 19440 ADD: Private car trips can only be made where there is a road connexion between the origin and the destination. The approximate average speed of that connexion is measured and used when determining whether a private car trip is made. CHANGE: If "quick_city_growth" is enabled, the renovation percentage is automatically reduced by a factor of 3. CHANGE: Refunds are based on 1.5x not 2x of the straight line journey distance.

+++ dataobj/einstellungen.h
@@ -182,7 +182,7 @@ private:
 	// The longest time that a passenger is
 	// prepared to wait for transport.
 	// @author: jamespetts
-	uint16 passenger_max_wait;
+	uint32 passenger_max_wait;
 
 	uint8 max_rerouting_interval_months;
